// src/app/api/model/userUserCreateReq.ts
/**
 * 
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface UserUserCreateReq { 
    email?: string;
    facultyId?: number;
    name?: string;
    password?: string;
    role?: UserUserCreateReq.RoleEnum;
}
export namespace UserUserCreateReq {
    export type RoleEnum = 'admin' | 'marketing_manager' | 'marketing_coordinator' | 'student' | 'guest';
    export const RoleEnum = {
        Admin: 'admin' as RoleEnum,
        MarketingManager: 'marketing_manager' as RoleEnum,
        MarketingCoordinator: 'marketing_coordinator' as RoleEnum,
        Student: 'student' as RoleEnum,
        Guest: 'guest' as RoleEnum
    };
}